# docs/openapi/ngsi-ld-api.yaml
openapi: 3.1.0
info:
  title: NGSI-LD OAS
  version: 1.8.1
  description: OpenAPI Specification for NGSI-LD API.
  contact:
    url: https://www.etsi.org/committee/cim
  license:
    name: BSD-3-Clause
    url: https://forge.etsi.org/legal-matters
...
servers:
  - url: '{protocol}://{hostname}:{port}/ngsi-ld/v1'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      hostname:
        default: localhost
      port:
        default: '443'
paths:
...
  /entityMap/{entityMapId}:
    get:
      tags:
        - Entity Map Operations
      summary: |
        EntityMap Retrieval by id 
      description:  |
        5.14.1 Retrieve EntityMap.

        With this operation a client can obtain a cached EntityMap which is currently stored in the broker's internal storage, or memory.
      operationId: retrieveEntityMap
      parameters:
        # Path params
        - $ref: '#/components/parameters/Path.entityMapId'
        # Request headers
        - $ref: '#/components/parameters/Headers.Link'
        - $ref: '#/components/parameters/Headers.ngsildTenant'
      responses:
        '200':
          description: |
            A response body containing the JSON-LD representation of the target entity.
          headers:
            NGSILD-Tenant:
              $ref: '#/components/headers/NGSILD-Tenant'
            NGSILD-Warning:
              $ref: '#/components/headers/NGSILD-Warning'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EntityMap'
                  - required:
                    - entityMap
                    - linkedMaps
            application/json+ld:
              schema:
                allOf:
                  - type: object
                    properties:
                      '@context':
                        $ref:  '#/components/schemas/LdContext'
                  - $ref: '#/components/schemas/EntityMap'
                  - required:
                    - entityMap
                    - linkedMaps
                    - '@context'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Entity Map Operations
      summary: |
        EntityMap Update by id
      description: |
        5.14.2 Update EntityMap.

        This operation allows performing a partial update on an NGSI-LD EntityMap. A partial update only changes the
        elements provided in the EntityMap, leaving the rest as they are.
      operationId: updateEntityMap
      parameters:
        # Path params
        - $ref: '#/components/parameters/Path.entityMapId'
        # Request headers
        - $ref: '#/components/parameters/Headers.Link'
        - $ref: '#/components/parameters/Headers.ngsildTenant'
      requestBody:
        description: |
          Payload body in the request contains a JSON-LD object which represents the context source registration that is to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityMap'
          application/json+ld:
            schema:
              allOf:
                - type: object
                  properties:
                    '@context':
                      $ref:  '#/components/schemas/LdContext'
                - $ref: '#/components/schemas/EntityMap'
                - required:
                  - '@context'
      responses:
        '204':
          description: |
            The EntityMap was updated successfully.
          headers:
            NGSILD-Tenant:
              $ref: '#/components/headers/NGSILD-Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Entity Map Operations
      summary: |
        EntityMap Deletion by id 
      description: |
        5.14.3 Delete EntityMap.

        This operation allows deleting an NGSI-LD EntityMap.
      operationId: deleteEntityMap
      parameters:
        # Path params
        - $ref: '#/components/parameters/Path.entityMapId'
        # Request headers
        - $ref: '#/components/parameters/Headers.Link'
        - $ref: '#/components/parameters/Headers.ngsildTenant'
      responses:
        '204':
          description: |
            No Content.
          headers:
            NGSILD-Tenant:
              $ref: '#/components/headers/NGSILD-Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  headers:
...
    NGSILD-Tenant:
      description: |
        6.3.14 Tenant specification. The tenant to which the NGSI-LD HTTP operation is targeted.
      schema:
        type: string
    NGSILD-Warning:
      description: |
        6.3.17 Distributed Operations Caching and Timeout Behaviour
      schema:
        type: string
        enum:
          - '110 - Response is Stale'
          - '111 - Revalidation Failed'
          - '199 - Miscellaneous Warning'
          - '299 - Miscellaneous Persistent Warning'
      required: false
  parameters:
...
    Headers.Link:
      name: Link
      in: header
      description: |
        6.3.5 JSON-LD @context resolution

        In summary, from a developer's perspective, for POST, PATCH and PUT operations,
        if MIME type is "application/ld+json", then the associated @context shall be provided
        only as part of the request payload body. Likewise, if MIME type is "application/json",
        then the associated @context shall be provided only by using the JSON-LD Link header.
        No mixes are allowed, i.e. mixing options shall result in HTTP response errors.
        Implementations should provide descriptive error messages when these situations arise.

        In contrast, GET and DELETE operations always take their input @context from the JSON-LD Link Header.
      explode: true
      schema:
        type: string
        format: uri
    Headers.ngsildTenant:
      name: NGSILD-Tenant
      in: header
      description: |
        6.3.14 Tenant specification. The tenant to which the NGSI-LD HTTP operation is targeted.
      explode: true
      schema:
        type: string
...
    Path.entityMapId:
      name: entityMapId
      in: path
      description: Id (URI) of the EntityMap to be retrieved, updated or deleted.
      schema:
        $ref: '#/components/schemas/Path'
      required: true
...
  schemas:
...
    LdContext:
      description: |
        5.2.3 JSON-LD @context

        When encoding NGSI-LD Entities, Context Source Registrations, Subscriptions and Notifications,
        as pure JSON-LD (MIME type "application/ld+json"), a proper @context shall be included
        as a special member of the corresponding JSON-LD Object.
      oneOf:
        - type: string
          format: uri
        - type: object
        - type: array
          items:
            oneOf:
              - type: string
                format: uri
              - type: object
...
    Path:
      description: Common schema for URI data type in path parameters.
      type: string
      format: uri
    ######
    # "ProblemDetails" schema from ETSI NFV OAS (BEGIN)
    #
    ProblemDetails:
      description: |
        The definition of the general "ProblemDetails" data structure from
        IETF RFC 7807 is reproduced inthis structure. Compared to the
        general framework defined in IETF RFC 7807, the "status" and
        "detail" attributes are mandated to be included by the present document,
        to ensure that the response contains additional textual information about
        an error. IETF RFC 7807 foresees extensibility of the
        "ProblemDetails" type. It is possible that particular APIs in the present
        document, or particular implementations, define extensions to define
        additional attributes that provide more information about the error.
        The description column only provides some explanation of the meaning to
        Facilitate understanding of the design. For a full description, see
        IETF RFC 7807.
      type: object
      required:
        - status
        - detail
      properties:
        type:
          description: |
            A URI reference according to IETF RFC 3986 that identifies the
            problem type. It is encouraged that the URI provides human-readable
            documentation for the problem (e.g. using HTML) when dereferenced.
            When this member is not present, its value is assumed to be
            "about:blank".
          type: string
          format: URI
        title:
          description: |
            A short, human-readable summary of the problem type. It should not
            change from occurrence to occurrence of the problem, except for
            purposes of localization. If type is given and other than
            "about:blank", this attribute shall also be provided.
            A short, human-readable summary of the problem
            type.  It SHOULD NOT change from occurrence to occurrence of the
            problem, except for purposes of localization (e.g., using
            proactive content negotiation; see [RFC7231], Section 3.4).
          type: string
        status:
          description: |
            The HTTP status code for this occurrence of the problem.
            The HTTP status code ([RFC7231], Section 6) generated by the origin
            server for this occurrence of the problem.
          type: integer
        detail:
          description: |
            A human-readable explanation specific to this occurrence of the
            problem.
            
            *InvalidRequest: The request associated to the operation is 
            syntactically invalid or includes wrong content. If an HTTP request 
            for an operation contains parameters that are incompatible with the 
            operation, or it contains values of the "options" parameter that are 
            not supported by the operation, an HTTP error response of type 
            InvalidRequest should be returned.  
          type: string
        instance:
          description: |
            A URI reference that identifies the specific occurrence of the
            problem. It may yield further information if dereferenced.
          type: string
          format: URI
    #
    # "ProblemDetails" schema from ETSI NFV OAS (END)
    ######
...
    EntityMap:
      description: |
        5.2.39 EntityMap.
      type: object
      properties:
        id:
          description: |
            EntityMap id.
          type: string
          format: uri
        type:
          description: |
            Node type.
          type: string
          enum:
            - EntityMap
          default: EntityMap
        expiresAt:
          description: |
            Expiration date for the EntityMap.
          type: string
          format: date-time
        entityMap:
          description: |
            System generated mapping of Entities to CSourceRegistrations.
            
            A set of key-value pairs whose keys shall be strings representing 
            Entity ids and whose values shall be an array holding every 
            CSourceRegistration id which is relevant to the ongoing Context 
            Information Consumption request (see clause 4.21). 
            
            The key "@none" shall be used to refer to an Entity that is held locally.
          type: object
          readOnly: true
        linkedMaps:
          description: |
            System generated mapping of Context CSourceRegistrations to a URI 
            indicating which EntityMaps was used by the Context Source.

            A set of key-value pairs whose keys shall be strings representing 
            CSourceRegistration ids which are relevant to the ongoing Context 
            Information request and whose values shall represent the associated 
            EntityMap id used by the ContextSource.
          type: object
          readOnly: true
      required:
        - type
        - expiresAt
...
  responses:
    BadRequest:
      description: |
        It is used to indicate that the request or its content is incorrect,
        see clause 6.3.2. In the returned ProblemDetails structure, the "detail"
        attribute should convey more information about the error.
      headers:
        NGSILD-Tenant:
          $ref: '#/components/headers/NGSILD-Tenant'
        NGSILD-Warning:
          $ref: '#/components/headers/NGSILD-Warning'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    NotFound:
      description: |
        It is used when a client provided an entity identifier (URI) not known
        to the system, see clause 6.3.2.
      headers:
        NGSILD-Tenant:
          $ref: '#/components/headers/NGSILD-Tenant'
        NGSILD-Warning:
          $ref: '#/components/headers/NGSILD-Warning'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...